+++ lib/openmaize/plugs/confirm.ex
@@ -9,12 +9,36 @@ defmodule Openmaize.Confirm do
   the Openmaize.Signup module for details about creating the token.
   """
 
+  use Openmaize.Pipe
+
   import Comeonin.Tools
   import Ecto.Changeset
   import Openmaize.Report
   alias Openmaize.{Config, QueryTools, Signup}
 
   @doc """
+  Function to confirm a user's email address.
+
+  ## Options
+
+  There are five options:
+
+  * key_expires_after - the length, in minutes, that the token is valid for
+    * the default is 1440 minutes, or one day
+  * unique_id - the identifier in the query string, or the parameters
+    * the default is :email
+  * mail_function - the emailing function that you need to define
+  * redirects - if Openmaize should handle the redirects or not
+    * the default true
+  * query_function - the function to query the database
+    * if you are using Ecto, you will probably not need this
+
+  ## Examples
+
+  After importing Openmaize.Confirm, run the following command:
+
+      plug :confirm_email, [mail_function: &Mailer.send_receipt/1] when action in [:confirm_email]
+
   """
   def confirm_email(%Plug.Conn{params: %{"key" => key} = user_params} = conn, opts)
   when byte_size(key) == 32 do
@@ -23,10 +47,21 @@ defmodule Openmaize.Confirm do
   def confirm_email(conn, opts), do: invalid_link_error(conn, opts)
 
   @doc """
+  Function to authenticate a user when resetting the password.
+
+  See the documentation for `confirm_email` for details about the available
+  options.
+
+  ## Examples
+
+  After importing Openmaize.Confirm, run the following command:
+
+      plug :reset_password, [mail_function: &Mailer.send_receipt/1] when action in [:reset_password]
+
   """
-  def reset_password(%Plug.Conn{params: %{"key" => key, "password" => password} = user_params} = conn, opts)
+  def reset_password(%Plug.Conn{params: %{"user" =>
+                    %{"key" => key, "password" => password} = user_params}} = conn, opts)
   when byte_size(key) == 32 do
-    Signup.create_user(%{"password" => password}) # validate changeset first
     check_user_key(conn, user_params, key, password, get_opts(opts))
   end
   def reset_password(conn, opts), do: invalid_link_error(conn, opts)
@@ -42,14 +77,14 @@ defmodule Openmaize.Confirm do
   defp check_user_key(conn, user_params, key, password,
                       {key_expiry, uniq, mail_func, redirects, query_func}) do
     user_id = Map.get(user_params, to_string(uniq))
-    user_id
-    |> URI.decode_www_form
-    |> query_func.(uniq)
-    |> check_key(key, key_expiry * 60, password)
+    error_pipe(user_id
+               |> URI.decode_www_form
+               |> query_func.(uniq)
+               |> check_key(key, key_expiry * 60, password))
     |> finalize(conn, user_id, mail_func, redirects)
   end
 
-  defp check_key({:error, message}, _, _, _), do: IO.inspect message
+  defp check_key(nil, _, _, _), do: false
   defp check_key(user, key, valid_secs, :nopassword) do
     check_time(user.confirmation_sent_at, valid_secs) and
     secure_check(user.confirmation_token, key) and
@@ -73,6 +108,9 @@ defmodule Openmaize.Confirm do
   end
   defp finalize(false, conn, user_id, _, redirects) do
     put_message(conn, "logout", %{"error" => "Confirmation for #{user_id} failed"}, redirects)
+  end
+  defp finalize(nil, conn, _, _, redirects) do
+    put_message(conn, "logout", %{"error" => "Confirmation failed"}, redirects)
   end
 
   defp invalid_link_error(conn, opts) do